Reverted patch which avoid using private ConnectionInfo API from DefaultHttpForwardedHeaderHandler (will create a new PR for this).
