+++ reactor-netty-http/src/test/java/reactor/netty/http/server/ConnectionInfoTests.java
@@ -15,6 +15,7 @@
  */
 package reactor.netty.http.server;
 
+import java.net.InetSocketAddress;
 import java.nio.charset.Charset;
 import java.security.cert.CertificateException;
 import java.time.Duration;
@@ -46,6 +47,7 @@ import reactor.netty.Connection;
 import reactor.netty.NettyPipeline;
 import reactor.netty.http.client.HttpClient;
 import reactor.netty.tcp.TcpClient;
+import reactor.netty.transport.AddressUtils;
 import reactor.test.StepVerifier;
 import reactor.util.annotation.Nullable;
 
@@ -60,7 +62,7 @@ import static reactor.netty.http.server.ConnectionInfo.DEFAULT_HTTP_PORT;
  *
  * @author Brian Clozel
  */
-public class ConnectionInfoTests extends BaseHttpTest {
+class ConnectionInfoTests extends BaseHttpTest {
 
 	static SelfSignedCertificate ssc;
 
@@ -432,6 +434,27 @@ public class ConnectionInfoTests extends BaseHttpTest {
 	}
 
 	@Test
+	void customForwardedHandlerForMultipleHost() {
+		testClientRequest(
+				clientRequestHeaders ->
+					clientRequestHeaders.add("X-Forwarded-Host", "a.example.com,b.example.com"),
+				serverRequest -> {
+					Assertions.assertThat(serverRequest.hostAddress().getHostString()).isEqualTo("b.example.com");
+					Assertions.assertThat(serverRequest.hostName()).isEqualTo("b.example.com");
+				},
+				(connectionInfo, request) -> {
+					String hostHeader = request.headers().get(DefaultHttpForwardedHeaderHandler.X_FORWARDED_HOST_HEADER);
+					if (hostHeader != null) {
+						InetSocketAddress hostAddress = AddressUtils.createUnresolved(
+								hostHeader.split(",", 2)[1].trim(),
+								connectionInfo.getHostAddress().getPort());
+						connectionInfo = connectionInfo.withHostAddress(hostAddress);
+					}
+					return connectionInfo;
+				});
+	}
+
+	@Test
 	void proxyProtocolOn() throws InterruptedException {
 		String remoteAddress = "202.112.144.236";
 		ArrayBlockingQueue<String> resultQueue = new ArrayBlockingQueue<>(1);
@@ -698,7 +721,7 @@ public class ConnectionInfoTests extends BaseHttpTest {
 		testClientRequest(clientRequestHeadersConsumer, serverRequestConsumer, null, Function.identity(), Function.identity(), false);
 	}
 
-	protected void testClientRequest(Consumer<HttpHeaders> clientRequestHeadersConsumer,
+	private void testClientRequest(Consumer<HttpHeaders> clientRequestHeadersConsumer,
 			Consumer<HttpServerRequest> serverRequestConsumer,
 			BiFunction<ConnectionInfo, HttpRequest, ConnectionInfo> forwardedHeaderHandler) {
 		testClientRequest(clientRequestHeadersConsumer, serverRequestConsumer, forwardedHeaderHandler, Function.identity(), Function.identity(), false);
